+++ src/test/java/io/github/dlbbld/ashlarchess/test/pgn/setup/PgnTestCaseCatalog.java
@@ -83,7 +83,7 @@ public class PgnTestCaseCatalog {
       case CHA_LICHESS_QUICK_DEPTH_ABOVE_FOUR_WINNABLE_FOR_FLAGGING_WITH_HELPMATE -> createTestCasesChaLichessQuickNotDepthThreeHelpmate();
       case CHA_LICHESS_QUICK_DEPTH_THREE -> createTestCasesChaLichessQuickDepthThree();
       case CHA_LICHESS_QUICK_DEPTH_FOUR -> createTestCasesChaLichessQuickDepthFour();
-      case CHA_AMBRONA -> createTestCasesChaAmbrona();
+      case CHA_CHASOLVER_EXCEPTIONS -> createTestCasesChaChasolverExceptions();
       case MAX_MOVES -> createTestCasesLongestPossible();
       case MAX_SAME_PIECE_PROMOTION_WHITE -> createTestCasesMaxSamePiecePromotionWhite();
       case MAX_SAME_PIECE_PROMOTION_BLACK -> createTestCasesMaxSamePiecePromotionBlack();
@@ -2609,34 +2609,12 @@ public class PgnTestCaseCatalog {
     return new PgnTestCaseList(PgnTest.CHA_BASIC_HELPMATE_EXISTENCE_THEOREM, list);
   }
 
-  private static PgnTestCaseList createTestCasesChaAmbrona() {
+  private static PgnTestCaseList createTestCasesChaChasolverExceptions() {
     final List<PgnFen> list = new ArrayList<>();
 
-    list.add(new PgnFen("ambrona_01.pgn", "5brk/4p1p1/3pP1P1/1B1P2p1/3p2p1/3P4/4K1P1/8 w - - 10 100"));
-    list.add(new PgnFen("ambrona_02.pgn", "8/1p4p1/1Pp3p1/k1P3p1/1pP3Pb/1P4p1/6P1/7K w - - 10 100"));
-    list.add(new PgnFen("ambrona_03.pgn", "8/1k5B/7b/8/1p1p1p1p/1PpP1P1P/2P3K1/N3b3 b - - 10 100"));
-    list.add(new PgnFen("ambrona_04.pgn", "7b/1k5B/7b/8/1p1p1p1p/1PpP1P1P/2P3K1/N7 b - - 10 100"));
-    list.add(new PgnFen("ambrona_04_proof_game.pgn", "7b/1k5B/7b/8/1p1p1p1p/1PpP1P1P/2P1P1K1/N7 b - - 6 56"));
-    list.add(new PgnFen("ambrona_05.pgn", "4K3/8/8/8/8/p1p2p1p/P1pppp1P/bnrqkrnb b - - 10 100"));
-    list.add(new PgnFen("ambrona_06.pgn", "k1bK4/1p1p4/1PpPp3/2P1Pp2/2p1pP2/2p1P3/2P5/8 w - - 10 100"));
-    list.add(new PgnFen("ambrona_07.pgn", "Bb2kb2/bKp1p1p1/1pP1P1P1/pP6/6P1/P7/8/8 b - - 10 100"));
-    list.add(new PgnFen("ambrona_08.pgn", "Bb2kb2/bKp1p1p1/1pP1P1P1/1P6/p5P1/P7/8/8 b - - 10 100"));
+    list.add(new PgnFen("chasolver_node_limit_exception.pgn", "1Bb5/1p6/pPp3k1/2Pp3p/P2PpBpP/4P1P1/5K2/8 w - - 8 32"));
 
-    // cannot use as position illegal (too many white light-squared bishops)
-    // nineth example from Ambrona, file name would become as below, but no such file as not used
-    // ae_09.pgn
-
-    list.add(new PgnFen("ambrona_10.pgn", "7k/8/1p6/1Pp5/2Pp4/pB1Pp1p1/P1B1P1P1/3B2K1 b - - 10 100"));
-
-    list.add(new PgnFen("ambrona_11_lichess_FKr42ZRT.pgn", "8/8/7p/5p1P/5p1K/5Pp1/6P1/5kb1 b - - 13 63"));
-    list.add(new PgnFen("ambrona_12_lichess_bKHPqNEw.pgn", "1k6/1P5p/BP3p2/1P6/8/8/5PKP/8 b - - 0 41"));
-    list.add(new PgnFen("ambrona_13_lichess_OawUhnkq.pgn", "5r1k/6P1/7K/5q2/8/8/8/8 b - - 0 51"));
-    list.add(new PgnFen("ambrona_14.pgn", "k7/Q6r/2b5/1pBp1p1p/1P1P1P1P/KP6/1P6/8 b - - 10 100"));
-    list.add(new PgnFen("ambrona_15_lichess_QRvIMh3z.pgn", "2b5/1p6/pPp3k1/2Pp3p/P2PpBpP/4P1P1/5K2/8 b - - 46 59"));
-    list.add(new PgnFen("ambrona_16.pgn", "1k6/p1p1p1p1/P1P1P1P1/p1p1p1p1/8/8/P1P1P1P1/4K3 w - - 10 100"));
-    list.add(new PgnFen("ambrona_17.pgn", "rnb1b3/pk1p4/p1pPp1p1/P1P1P1P1/RBP5/P7/5B2/7K w - - 10 100"));
-
-    return new PgnTestCaseList(PgnTest.CHA_AMBRONA, list);
+    return new PgnTestCaseList(PgnTest.CHA_CHASOLVER_EXCEPTIONS, list);
   }
 
   // Fixtures the geometric pawn-wall classifier accepts as YES: chain spans file a to h, both kings on opposite
@@ -2710,7 +2688,7 @@ public class PgnTestCaseCatalog {
     list.add(new PgnFen("pawn_wall_norgaard_additional_own_pawns_not_marched_up.pgn",
         "6k1/1p1p1p1p/1P1P1P1P/8/8/4K3/1P1P1P1P/8 w - - 0 39"));
 
-    // TODO today's Ambrona website (full or quick) sees unwinnable for both sides
+    // the example with different outcome in C++ CHA
     list.add(new PgnFen("pawn_wall_norgaard_additional_own_pawns_not_marched_up_with_opponent_pawns_between.pgn",
         "1k6/p1p1p1p1/P1P1P1P1/p1p1p1p1/8/8/P1P1P1P1/4K3 w - - 0 34"));